+++ Navigation/GalleryViewContext.cs
@@ -1,5 +1,6 @@
 using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI;
+using StardewGallery.Appearance;
 
 namespace StardewGallery;
 
@@ -9,6 +10,7 @@ internal sealed record GalleryTextures(Texture2D Home, Texture2D Album, Texture2
 internal sealed record GalleryViewContext(GalleryCatalog Catalog, ITranslationHelper I18n, GalleryTextures Textures,
     GalleryPhotos Photos, IGalleryNavigation Navigation, Func<bool> IsUnlocked, Func<GalleryEvent, ReplayAccess> ReplayAccess)
 {
+    internal ICharacterAppearance Appearance { get; init; } = null!;
     internal EventNameStore? Names { get; init; }
     internal GalleryLocationNames Locations { get; } = new(I18n);
     internal GalleryCharacterNames Characters { get; } = new(Catalog, I18n);